requirements.txt
Flask==1.1.2
google-cloud-storage==1.29.0
gunicorn==20.0.4; python_version > '3.0'
nltk==3.5
matplotlib==3.3.0
wordcloud==1.7.0